fix: don't save the last checkpoint as a symlink
This option doesn't seem to work like I'd expect, when also saving the
best models according to some metric.  I expected to get
descriptively-named real files for the metric and the last epoch, and a
fixed-name symlink to the latter.

In fact, I seem to only get a symlink to the metric checkpoint and no
last-epoch checkpoint.  Basically, the "symlink" option means "don't
create a new checkpoint, just make a link to whatever existing
checkpoint is most recent".  I read the code and think this is what's
happening, but it's a little hard to be sure.

By switching `save_last='link'` to `save_link=True`, I always get two
checkpoints: one for the best metric and one for the last epoch.